hr-chatbot/4-eval-driven-development/requirements.txt
arize-phoenix
arize-phoenix-otel
openai
openinference-instrumentation-openai
python-dotenv
pytest